requirements.txt
graphene-django
faker
drf-yasg
djangorestframework